chore(sentinel): cleaned up macro code

+++ macros/byteme/src/view.rs
@@ -13,11 +13,11 @@ pub enum ServerGreetingMode {
 #[derive(ByteMe)]
 pub struct ServerGreetingFrame {
   pub unused: [u8; 12],
-  #[byte_me(u8)]
+  #[byte_me(u32)]
   pub mode: ServerGreetingMode,
   pub challenge: [u8; 16],
   pub salt: [u8; 16],
-  pub count: u32,
+  pub count: u128,
   pub mbz: [u8; 12],
 }
